Found a small bug adding new note
Adding a new note could be achieved by clicking on the whole status bar (AppLoggedIn level) instead of only clicking on the + icon. There is a small bug where after adding a new note, when the App udpates and shows the list of notes, the button is green then turns back to normal.

files modified:
AddNewNote.js
AppLoggedIn.js